In an effort to reduce warnings and errors caused by new compilers, the types for vrpn_int32 and vrpn_uint32 in _WIN32 (all compilers, including pc_cygwin and VC++) has been changed to int, instead of long.
This BREAKS compatibility with WINDOWS 3.1. Hopefully this is not an
issue.

+++ vrpn_Types.h
@@ -109,8 +109,8 @@ typedef  char            vrpn_int8;
 typedef  unsigned char   vrpn_uint8;
 typedef  short           vrpn_int16;
 typedef  unsigned short  vrpn_uint16;
-typedef  long            vrpn_int32;
-typedef  unsigned long   vrpn_uint32;
+typedef  int             vrpn_int32;
+typedef  unsigned int   vrpn_uint32;
 typedef  float           vrpn_float32;
 typedef  double          vrpn_float64;
 #endif
